added base module and new base types for models, aesthetic fixes in vector model
the new base types allow to have models with fit only, predict only or both
based VectorModel off FitPredictModel and renamed _log to log
